requirements.txt
fastapi==0.109.0
langchain==0.1.3
langchain_community==0.0.15
langchain_core==0.1.15
langchain_openai==0.0.3
langserve==0.0.39
protobuf==4.25.2
pydantic==2.5.3
python-dotenv==1.0.1
Requests==2.31.0
uvicorn==0.27.0
